:art: rename everything related to subscription to recurrence product
