+++ Controller/Adminhtml/RecurrenceProducts/Delete.php
@@ -1,24 +1,19 @@
 <?php
 
-namespace MundiPagg\MundiPagg\Controller\Adminhtml\Subscriptions;
+namespace MundiPagg\MundiPagg\Controller\Adminhtml\RecurrenceProducts;
 
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
+use Magento\Framework\Message\Factory;
 use Magento\Framework\Registry;
 use Magento\Framework\View\Result\PageFactory;
-use Mundipagg\Core\Recurrence\Aggregates\ProductSubscription;
 use Mundipagg\Core\Recurrence\Services\ProductSubscriptionService;
 use MundiPagg\MundiPagg\Concrete\Magento2CoreSetup;
 use MundiPagg\MundiPagg\Model\ProductsSubscriptionFactory;
 
-class Create extends Action
+class Delete extends Action
 {
     protected $resultPageFactory;
-
-    /**
-     * @var ProductsSubscriptionFactory
-     */
-    protected $productsSubscriptionFactory;
 
     /**
      * @var Registry
@@ -30,17 +25,18 @@ class Create extends Action
      *
      * @param \Magento\Backend\App\Action\Context $context
      * @param \Magento\Framework\View\Result\PageFactory $resultPageFactory
+     * @throws \Exception
      */
     public function __construct(
         \Magento\Backend\App\Action\Context $context,
         \Magento\Framework\View\Result\PageFactory $resultPageFactory,
         Registry $coreRegistry,
-        ProductsSubscriptionFactory $productsSubscriptionFactory
+        Factory $messageFactory
     )
     {
         $this->resultPageFactory = $resultPageFactory;
-        $this->productsSubscriptionFactory = $productsSubscriptionFactory;
         $this->coreRegistry = $coreRegistry;
+        $this->messageFactory = $messageFactory;
         Magento2CoreSetup::bootstrap();
 
         parent::__construct($context);
@@ -60,20 +56,19 @@ class Create extends Action
             $productData = $productSubscriptionService->findById($productId);
 
             if (!$productData || !$productData->getId()) {
-                $this->messageManager->addError(__('Product subscription not exist.'));
-                $this->_redirect('mundipagg_mundipagg/subscriptions/index');
+                $message = $this->messageFactory->create('error', __('Product subscription not exist.'));
+                $this->messageManager->addErrorMessage($message);
+                $this->_redirect('mundipagg_mundipagg/recurrenceproducts/index');
                 return;
             }
-
-            $this->coreRegistry->register('subscription_data', json_encode($productData));
         }
-        $this->coreRegistry->register('recurrence_type', ProductSubscription::RECURRENCE_TYPE);
 
-        $title = $productId ? __('Edit Recurrence Product') : __('Create Recurrence Product');
+        $productSubscriptionService->delete($productId);
 
-        $resultPage = $this->resultPageFactory->create();
-        $resultPage->getConfig()->getTitle()->prepend($title);
+        $message = $this->messageFactory->create('success', _("Product subscription deleted."));
+        $this->messageManager->addMessage($message);
 
-        return $resultPage;
+        $this->_redirect('mundipagg_mundipagg/recurrenceproducts/index');
+        return;
     }
 }
