+++ i18n/pt_BR.csv
@@ -371,3 +371,4 @@
 "year", "ano"
 "subscription", "assinatura"
 "plan", "plano"
+"active", "ativo"
