+++ view/adminhtml/layout/mundipagg_mundipagg_recurrenceproducts_index.xml
@@ -3,7 +3,7 @@
 	<update handle="styles"/>
 	<body>
 		<referenceContainer name="content">
-			<uiComponent name="mundipagg_mundipagg_subscriptions_listing"/>
+			<uiComponent name="mundipagg_mundipagg_recurrenceproducts_listing"/>
 		</referenceContainer>
 	</body>
 </page>
