+++ view/adminhtml/layout/sales_order_view.xml
