+++ view/adminhtml/templates/recurrence/subscriptions/create.phtml
@@ -15,7 +15,7 @@ $quantityOptionCicle = 4;
                     </button>
                     <button id="back" title="Back" type="button"
                             class="action- scalable back"
-                            onclick="location.href = '<?php echo $block->getUrl('mundipagg_mundipagg/subscriptions/index') ?>'"
+                            onclick="location.href = '<?php echo $block->getUrl('mundipagg_mundipagg/recurrenceproducts/index') ?>'"
                             data-ui-id="back-button">
                         <span><?php echo __('Back') ?></span>
                     </button>
