+++ view/adminhtml/ui_component/mundipagg_mundipagg_recurrenceproducts_listing.xml
@@ -1,8 +1,8 @@
 <listing xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Ui:etc/ui_configuration.xsd">
 	<argument name="data" xsi:type="array">
 		<item name="js_config" xsi:type="array">
-			<item name="provider" xsi:type="string">mundipagg_mundipagg_subscriptions_listing.mundipagg_mundipagg_subscriptions_listing_data_source</item>
-			<item name="deps" xsi:type="string">mundipagg_mundipagg_subscriptions_listing.mundipagg_mundipagg_subscriptions_listing_data_source</item>
+			<item name="provider" xsi:type="string">mundipagg_mundipagg_recurrenceproducts_listing.mundipagg_mundipagg_recurrenceproducts_listing_data_source</item>
+			<item name="deps" xsi:type="string">mundipagg_mundipagg_recurrenceproducts_listing.mundipagg_mundipagg_recurrenceproducts_listing_data_source</item>
 		</item>
 		<item name="spinner" xsi:type="string">spinner_columns</item>
 		<item name="buttons" xsi:type="array">
@@ -29,7 +29,7 @@
 	<dataSource name="nameOfDataSource">
 		<argument name="dataProvider" xsi:type="configurableObject">
 			<argument name="class" xsi:type="string">Magento\Framework\View\Element\UiComponent\DataProvider\DataProvider</argument>
-			<argument name="name" xsi:type="string">mundipagg_mundipagg_subscriptions_listing_data_source</argument>
+			<argument name="name" xsi:type="string">mundipagg_mundipagg_recurrenceproducts_listing_data_source</argument>
 			<argument name="primaryFieldName" xsi:type="string">id</argument>
 			<argument name="requestFieldName" xsi:type="string">id</argument>
 			<argument name="data" xsi:type="array">
